feat: implement validation of adapter configuration (#11)
### Description

This PR adds configuration validation in the init() function of the CCXT adapter. This validation process checks the adapter-specific configuration for any inconsistencies, omissions, or errors that could disrupt normal functioning. If any issues are found during the validation process, an error is thrown. 
These can be ignored and will be further modified, then can be fully reviewed in a future PR. 

### Other changes

- Added a class to validate and load the app config file

### Tested

- Added unit tests to ensure the validation process works as expected, with tests for both valid and invalid configurations

### Related issues

-  Closes #7 

### Backwards compatibility

- N/A